backend/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiohttp-retry==2.9.1
aiosignal==1.3.2
aiosqlite==0.21.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
bcrypt==3.2.0
cachetools==5.5.2
certifi==2025.4.26
cffi==1.17.1
charset-normalizer==3.4.2
click==8.1.8
colorama==0.4.6
databases==0.9.0
dataclasses-json==0.6.7
docx2txt==0.9
ecdsa==0.19.1
faiss-cpu==1.11.0
fastapi==0.115.12
filelock==3.18.0
filetype==1.2.0
frozenlist==1.6.0
fsspec==2025.3.2
google-ai-generativelanguage==0.6.18
google-api-core==2.24.2
google-auth==2.40.2
googleapis-common-protos==1.70.0
greenlet==3.2.2
grpcio==1.72.0rc1
grpcio-status==1.72.0rc1
gTTS==2.5.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.31.2
idna==3.10
Jinja2==3.1.6
joblib==1.5.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.25
langchain-community==0.3.24
langchain-core==0.3.59
langchain-google-genai==2.1.4
langchain-text-splitters==0.3.8
langsmith==0.3.42
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.4.3
mypy_extensions==1.1.0
networkx==3.4.2
numpy==2.2.5
orjson==3.10.18
packaging==24.2
passlib==1.7.4
pillow==11.2.1
propcache==0.3.1
proto-plus==1.26.1
protobuf==6.31.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.22
pydantic==2.11.4
pydantic-settings==2.9.1
pydantic_core==2.33.2
pydub==0.25.1
PyJWT==2.10.1
PyMuPDF==1.25.5
pypdf==5.5.0
python-dotenv==1.1.0
python-jose==3.4.0
python-multipart==0.0.20
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rsa==4.9.1
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.3
sentence-transformers==4.1.0
setuptools==80.7.1
six==1.17.0
sniffio==1.3.1
SpeechRecognition==3.14.3
SQLAlchemy==2.0.41
standard-aifc==3.13.0
standard-chunk==3.13.0
starlette==0.46.2
sympy==1.14.0
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.7.0
tqdm==4.67.1
transformers==4.51.3
twilio==9.6.1
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.2
urllib3==2.4.0
uvicorn==0.34.2
yarl==1.20.0
zstandard==0.23.0
openai-whisper==20231117
torch>=2.0.0
--extra-index-url https://download.pytorch.org/whl/cpu


